Move the loading process of saturation config from dataset to metric
